Added cors to server, added name input into register form, added fetch
